refactor(evm-wallet-experiment): rename simulation.md→demo-local.md and interactive-*.mjs→demo-*.mjs
