+++ packages/evm-wallet-experiment/docker/run-interactive-compose.mjs
@@ -1,3 +1,3 @@
-import { runInteractiveCompose } from './interactive-compose-lib.mjs';
+import { runInteractiveCompose } from './demo-compose-lib.mjs';
 
 runInteractiveCompose(process.argv.slice(2));
