+++ packages/evm-wallet-experiment/package.json
@@ -68,8 +68,8 @@
     "docker:demo:up": "yarn docker:ensure-logs && node docker/run-interactive-compose.mjs up --build",
     "docker:demo:down": "node docker/run-interactive-compose.mjs down",
     "docker:setup:wallets": "yarn tsx test/e2e/docker/setup-wallets.ts",
-    "docker:demo:setup": "node docker/interactive-setup.mjs",
-    "docker:demo:reset-openclaw": "node docker/interactive-reset-openclaw.mjs",
+    "docker:demo:setup": "node docker/demo-setup.mjs",
+    "docker:demo:reset-openclaw": "node docker/demo-reset-openclaw.mjs",
     "docker:demo:attach:away": "node docker/attach.mjs away",
     "docker:demo:attach:home": "node docker/attach.mjs home",
     "docker:delegate": "node docker/delegate.mjs",
